Keycloak provider: Set optional configuration items default values to null instead of an empty string to avoid a PHP warning.
The PHP warning is as follows: `PHP Warning: file_get_contents(): Filename cannot be empty in [...]/vendor/stevenmaguire/oauth2-keycloak/src/Provider/Keycloak.php on line 242`.
This warning was thrown because the Keycloak provider tries to read the file at encryption_key_path when this option is present in the array of options, which is the case if the default value of this configuration item is an empty string.