Update Azure DI Config With Default Values
The update to add all of the Azure-specific configuration variables into
the tree builder did not set any defaults, so you have to configure each
and every option in the yaml file.

This updates #131 to include the proper default values as defined in the
TheNetworg/oauth2-azure Provider class. This way configuration can be as
minimal as possible, allowing you to simply overwrite default values
instead of defining each and every one.